Check every native definition against a declaration
Four findings, from pointing a strict warning set and clang-tidy at
native/ before deciding whether to keep either.

logsink.c never included logsink.h, so the three functions the FFI
binding is generated from were checked against nothing. Nothing has
drifted yet; the point is that nothing would have said so, and a
signature that drifts from its cdef fails as a wrong offset rather than
as a missing symbol. spirvcross.c had no header at all, which is why it
was the one module the same check could not be applied to, so it has one
now like every other.

The registry generator emits the accessor's declaration ahead of its
definition for the same reason: registry.c declares it too, from the
other side of a link, and neither copy was being held against the other.

TECS_BACKGROUND_BUDGET_NS multiplied in int. A quarter of a second fits
and anything past about two seconds does not, so raising the budget
would have wrapped it into one nobody could exceed rather than failing
to compile.

The rest is the two sites the warnings named: the sink's unused
userdata, and an enum reaching an int return through an implicit
narrowing that is now written down.

+++ native/host.c
@@ -67,8 +67,13 @@ typedef struct TecsHost TecsHost;
  * rather less, so this sits well inside both. It is not there to enforce the
  * platform's deadline, which the host cannot do; it is there to catch the one
  * mistake this path cannot afford, which is a game walking its world inside the
- * callback instead of flushing a checkpoint it prepared during a frame. */
-#define TECS_BACKGROUND_BUDGET_NS (250 * SDL_NS_PER_MS)
+ * callback instead of flushing a checkpoint it prepared during a frame.
+ *
+ * The count is written as `Uint64` because the product is nanoseconds and the
+ * multiplication would otherwise be done in `int`. A quarter of a second fits;
+ * anything above about two seconds does not, and would wrap into a budget
+ * nobody could exceed rather than failing to compile. */
+#define TECS_BACKGROUND_BUDGET_NS ((Uint64)250 * SDL_NS_PER_MS)
 
 /* One iteration's worth of events, and everything they point at.
  *
